Move image transformation methods to Utils class
