+++ app/src/main/java/social/entourage/android/tools/Utils.kt
@@ -1,12 +1,13 @@
 package social.entourage.android.tools
 
+import android.content.ContentResolver
 import android.content.Context
-import android.graphics.Bitmap
-import android.graphics.Canvas
-import android.graphics.Point
-import android.graphics.Typeface
+import android.graphics.*
 import android.graphics.drawable.Drawable
+import android.net.Uri
 import android.os.Build
+import android.os.Environment
+import android.provider.MediaStore
 import android.text.Html
 import android.text.SpannableStringBuilder
 import android.text.Spanned
@@ -20,6 +21,10 @@ import android.view.WindowManager
 import com.google.android.gms.maps.model.BitmapDescriptor
 import com.google.android.gms.maps.model.BitmapDescriptorFactory
 import social.entourage.android.R
+import java.io.File
+import java.io.FileOutputStream
+import java.io.IOException
+import java.text.SimpleDateFormat
 import java.util.*
 import kotlin.math.max
 
@@ -183,5 +188,36 @@ object Utils {
             }
         }
         return builder
+    }
+
+    fun getBitmapFromUri(uri: Uri, contentResolver: ContentResolver?): Bitmap {
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P && contentResolver != null)
+            ImageDecoder.decodeBitmap(ImageDecoder.createSource(contentResolver, uri))
+        else
+            MediaStore.Images.Media.getBitmap(contentResolver, uri)
+    }
+
+    fun saveBitmapToFile(bitmap: Bitmap, file: File?): File {
+        val photoFile = file ?: createImageFile()
+
+        FileOutputStream(photoFile).use { out ->
+            bitmap.compress(Bitmap.CompressFormat.PNG, 100, out)
+        }
+
+        return photoFile
+    }
+
+    @Throws(IOException::class)
+    private fun createImageFile(): File {
+        // Create an image file name
+        val timeStamp = SimpleDateFormat("yyyyMMdd_HHmmss", Locale.FRANCE).format(Date())
+        val imageFileName = "ENTOURAGE_CROP_" + timeStamp + "_"
+        val storageDir = Environment.getExternalStoragePublicDirectory(
+                Environment.DIRECTORY_PICTURES)
+        return File.createTempFile(
+                imageFileName,
+                ".jpg",
+                storageDir
+        )
     }
 }
